Splitting and moving HTML tools to a directory
