+++ src/tools/html/get-plaintext.ts
@@ -1,0 +1,5 @@
+import { htmlToText, HtmlToTextOptions } from 'html-to-text';
+
+export function getPlainText(html: string, options?: HtmlToTextOptions): string {
+  return htmlToText(html, options);
+}
