+++ src/tools/html/parse-with-cheerio.ts
@@ -1,0 +1,8 @@
+import * as cheerio from 'cheerio';
+
+export function parseWithCheerio(
+  html: string | Buffer,
+  options?: Parameters<typeof cheerio.load>[1]
+ ): cheerio.Root {
+  return cheerio.load(html, options);
+}
